+++ src/commands/schemas.ts
@@ -5,6 +5,8 @@ import { Logger } from "../logging";
 import { Schema } from "../models/schema";
 import { SchemaRegistryCluster } from "../models/schemaRegistry";
 import { getSchemasViewProvider } from "../viewProviders/schemas";
+import { KafkaTopic } from "../models/topic";
+import { ResourceManager } from "../storage/resourceManager";
 
 const logger = new Logger("commands.schemas");
 
@@ -55,6 +57,44 @@ function uploadVersionCommand(item: any) {
   );
 }
 
+async function openLatestSchemasCommand(topic: KafkaTopic) {
+  let highestVersionedSchemas: Schema[] | null = null;
+
+  try {
+    highestVersionedSchemas = await getLatestSchemasForTopic(topic);
+  } catch (e) {
+    if (e instanceof CannotLoadSchemasError) {
+      logger.error(e.message);
+      vscode.window.showErrorMessage(e.message);
+      return;
+    } else {
+      throw e;
+    }
+  }
+
+  // Make a nice message to show in the progress bar, albeit short lived.
+  const schemaSubjectVersionList = highestVersionedSchemas
+    .map((s) => `${s.subject} (${s.version})`)
+    .join(", ");
+  const maybe_ess = highestVersionedSchemas.length > 1 ? "s" : "";
+  const message = `Opening latest schema${maybe_ess} for topic "${topic.name}": "${schemaSubjectVersionList}"`;
+
+  logger.info(message);
+
+  await vscode.window.withProgress(
+    {
+      location: vscode.ProgressLocation.Notification,
+      title: message,
+    },
+    async () => {
+      const promises = highestVersionedSchemas.map((schema) => {
+        loadOrCreateSchemaViewer(schema);
+      });
+      await Promise.all(promises);
+    },
+  );
+}
+
 export function registerSchemaCommands(): vscode.Disposable[] {
   return [
     registerCommandWithLogging("confluent.schemaViewer.refresh", refreshCommand),
@@ -62,6 +102,7 @@ export function registerSchemaCommands(): vscode.Disposable[] {
     registerCommandWithLogging("confluent.schemaViewer.uploadVersion", uploadVersionCommand),
     registerCommandWithLogging("confluent.schemaViewer.viewLocally", viewLocallyCommand),
     registerCommandWithLogging("confluent.schemas.copySchemaRegistryId", copySchemaRegistryId),
+    registerCommandWithLogging("confluent.topics.openlatestschemas", openLatestSchemasCommand),
   ];
 }
 
@@ -78,4 +119,70 @@ async function loadOrCreateSchemaViewer(schema: Schema) {
   // but they don't show up to the user unless they look at the "Window" output channel.
   vscode.languages.setTextDocumentLanguage(textDoc.document, schema.fileExtension());
   return textDoc;
+}
+
+/**
+ * Get the highest versioned schema(s) related to a single topic from the schema registry
+ * as decided by TopicNameStrategy. May return two schemas if the topic has both key and value schemas.
+ */
+export async function getLatestSchemasForTopic(topic: KafkaTopic): Promise<Schema[]> {
+  // These two checks indicate programming errors, not a user or external system contents issues ...
+  if (!topic.hasSchema) {
+    throw new Error(`Asked to get schemas for topic "${topic.name}" believed to not have schemas.`);
+  }
+
+  // local topics, at time of writing, won't have any related schemas, 'cause we don't support any form
+  // of local schema registry (yet). But when supporting local schema registry will probably need a different
+  // way to get schemas than these ccloud-infected methods, so raise an error here as a reminder to revisit
+  // this code when local schema registry support is added.
+  if (topic.isLocalTopic()) {
+    throw new Error(
+      `Asked to get schemas for local topic "${topic.name}", but local topics should not have schemas.`,
+    );
+  }
+
+  const rm = ResourceManager.getInstance();
+
+  const schemaRegistry = await rm.getCCloudSchemaRegistryCluster(topic.environmentId!);
+  if (schemaRegistry === null) {
+    throw new CannotLoadSchemasError(
+      `Could not determine schema registry for topic "${topic.name}" believed to have related schemas.`,
+    );
+  }
+
+  const allSchemas = await rm.getSchemasForRegistry(schemaRegistry.id);
+
+  if (allSchemas === undefined || allSchemas.length === 0) {
+    throw new CannotLoadSchemasError(
+      `Schema registry "${schemaRegistry.id}" had no schemas, but we expected it to have some for topic "${topic.name}"`,
+    );
+  }
+
+  // Filter by TopicNameStrategy for this topic.
+  const topicSchemas = allSchemas.filter((schema) => schema.matchesTopicName(topic.name));
+
+  // Now make map of schema subject -> highest version'd schema for said subject
+  const nameToHighestVersion = new Map<string, Schema>();
+  for (const schema of topicSchemas) {
+    const existing = nameToHighestVersion.get(schema.subject);
+    if (existing === undefined || existing.version < schema.version) {
+      nameToHighestVersion.set(schema.subject, schema);
+    }
+  }
+
+  if (nameToHighestVersion.size === 0) {
+    throw new CannotLoadSchemasError(`No schemas found for topic "${topic.name}"!`);
+  }
+
+  // Return flattend values from the map, the list of highest-versioned schemas related to the topic.
+  return [...nameToHighestVersion.values()];
+}
+
+/** Raised when unexpectedly could not load schema(s) for a topic we previously believed
+ * had related schemas. Message will be informative and user-facing.
+ */
+export class CannotLoadSchemasError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
 }