+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this extension will be documented in this file.
 
 ## Unreleased
 
+- New context menu item "View Latest Schema Version(s)" to quickly open the highest versioned value
+  and / or key schemas for a CCloud topic according to TopicNameStrategy,
+  [issue #261](https://github.com/confluentinc/vscode/issues/261).
+
 ## 0.17.1
 
 - Hardened handling of empty bins in the histogram of the message viewer